+++ CHANGELOG.md
@@ -10,14 +10,22 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Breaking changes
 
+## 3.2.0 2020-06-24
+
+### Compatible changes
+
+- Add a `.geordi.yml` file to change multiple settings in the project and  `~/.config/geordi/global.yml` for global settings.
+- Add #91: Now there is an option to start cucumber without a VNC session. This is configured by the .geordi.yml file.
+- Deprecated the `.pt_project_id` file in favor of `.geordi.yml`.
+- Deprecated the `~/.gitpt` file in favor of `~/.config/geordi/global.yml`.
+- Fixed `git#staged_changes?` detection on Ruby < 2.5.
+
 
 ## 3.1.0 2020-06-03
 
 ### Compatible changes
 
 - Update security-update for improved workflow (#89): Deploy staging first and ask user, if application is still running. Then deploy other stages.
-
-### Breaking changes
 
 
 ## 3.0.3 2020-05-27
@@ -31,7 +39,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Compatible changes
 
-- Fix #95: Method change from `! *.include?` to `*.exclude?` was not valid as we do not have active support in Geordi. Affected commands where `geordi cucumber` and `geordi deploy`.
+- Fix #95: Method change from `! *.include?` to `*.exclude?` was not valid as we do not have active support in Geordi. Affected commands were `geordi cucumber` and `geordi deploy`.
 
 ### Breaking changes